Add test harness for frontend app.
