+++ package.json
@@ -31,10 +31,12 @@
   },
   "devDependencies": {
     "jest": "^22.4.3",
+    "keygrip": "^1.0.2",
     "nock": "^9.3.2",
     "nodemon": "^1.17.2",
     "smee-client": "^1.0.1",
     "standard": "^11.0.1",
+    "supertest": "^3.1.0",
     "testdouble": "^3.8.1",
     "testdouble-jest": "^2.0.0",
     "testdouble-nock": "^0.2.0"
